+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Add 404 fallback routing
 - Add about page
 - Add games page.
+- Add basic filesystem.
 
 ### Changed